requirements.txt
whisper.ai
NumPy